requirements.txt
pandas>=2.0
google-cloud-storage>=2.0
pyarrow>=14.0
python-dotenv>=1.0
tqdm>=4.0